+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - 使用 pm2 config 控管 @esbb48
 - 新增健康與遊戲專區的版位 @esbb48
 
+### Changed
+- 移除舊 NiceGame 的 DFP 廣告 @esbb48
+
 ## 1.0.34 - 2017-10-12
 ### Changed
 - 調整子網站的圖片變為多彩 @esbb48